Update upload result table
