+++ src/components/Upload/Upload.jsx
@@ -1,8 +1,9 @@
 import { TSTemplate } from "../../models/Templates";
-import WorkloadFactory from "../Workload/WorkloadFactory";
-import SearchFields from "../SearchFields/SearchFields";
 import { useContext, useState } from "react";
 import { FieldContext } from "../../context/FieldContext";
+import WorkloadFactory from "../Workload/WorkloadFactory";
+import SearchFields from "../SearchFields/SearchFields";
+import UploadResult from "./UploadResult";
 
 function Upload() {
     const { formState, handleTimeSeriesUpload } = useContext(FieldContext);
@@ -18,7 +19,7 @@ function Upload() {
                 body: JSON.stringify(formState),
             });
             const result = await entity.json();
-            setSubmissionResult(formState);
+            setSubmissionResult(result);
             console.log("Workload created: ", result);
         } catch (error) {
             console.error("Error creating workload: ", error);
@@ -44,29 +45,6 @@ function Upload() {
             window.URL.revokeObjectURL(url);
         };
         createTemplate(TSTemplate, "time-series");
-    };
-
-    const generateTable = () => {
-        const excludedKeys = ["timeSeries", "userDefinedFields"];
-
-        const configurePart = Object.entries(formState)
-            .filter(([key, value]) => !excludedKeys.includes(key))
-            .map(([key, value]) => (
-                <tr className="table-content" key={key}>
-                    <td className="pt-2 pb-2 td-key">{key}</td>
-                    <td className="pt-2 pb-2">{value}</td>
-                </tr>
-            ));
-
-        const userDefined = Object.keys(formState.userDefinedFields);
-        const userDefinedPart = userDefined.map((key) => (
-            <tr className="table-content" key={key}>
-                <td className="pt-2 pb-2 td-key">{key}</td>
-                <td className="pt-2 pb-2">{formState.userDefinedFields[key]}</td>
-            </tr>
-        ));
-
-        return [...configurePart, ...userDefinedPart];
     };
 
     return (
@@ -101,34 +79,7 @@ function Upload() {
                 <button className="btn btn-outline-primary col-md-5" onClick={(event) => handleSubmit(event)}>
                     Submit
                 </button>
-                {submissionResult && (
-                    <div className={`alert ${submissionResult.error ? "alert-danger" : "alert-success"}`} role="alert">
-                        {submissionResult.error ? (
-                            <div>
-                                <strong>{submissionResult.error}</strong>
-                                <br />
-                                <strong>Error Details:</strong> {submissionResult.details}
-                            </div>
-                        ) : (
-                            <div>
-                                <strong>Workload created successfully!</strong>
-                                {submissionResult && Object.keys(submissionResult).length > 0 ? (
-                                    <table className="table">
-                                        <thead>
-                                            <tr>
-                                                <th>Key</th>
-                                                <th>Value</th>
-                                            </tr>
-                                        </thead>
-                                        <tbody>{generateTable()}</tbody>
-                                    </table>
-                                ) : (
-                                    <p>Loading...</p>
-                                )}
-                            </div>
-                        )}
-                    </div>
-                )}
+                <UploadResult formState={formState} submissionResult={submissionResult} />
             </div>
         </div>
     );
